Fixing cycling dependency issues
