+++ src/components/project/projectHelpers.js
@@ -35,7 +35,6 @@ const projectFormInputs = () => {
   const projectDescription = input.input('description');
   const projectDate = input.dateInput('date');
   const submitBtn = btn.addBtn('Create Project');
-  // projectForm.addEventListener('submit', helpers.addNewProject);
   projectForm.innerHTML = projectTitle + projectDescription + projectDate;
   projectForm.appendChild(submitBtn);
 };
